requirements.txt
# OCR Corrector - Complete Requirements
# Install: python -m pip install -r requirements.txt

# Core Web Framework
streamlit>=1.28.0
streamlit-option-menu>=0.3.2
flask>=3.0.0
flask-cors>=4.0.0

# Environment & Configuration
python-dotenv>=1.0.0

# Image Processing
opencv-python>=4.8.0
pillow>=10.0.0
pdf2image>=1.16.0
pytesseract>=0.3.10
PyMuPDF>=1.23.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# PDF Processing
pdfplumber>=0.10.0
pypdf>=3.17.0

# OCR & Language
google-cloud-vision>=3.4.0
langdetect>=1.0.9
Levenshtein>=0.21.0

# Utilities
tqdm>=4.65.0
requests>=2.30.0
watchdog>=3.0.0

# ML/Deep Learning (if needed for model inference)
torch>=2.0.0
torchvision>=0.15.0

# Additional utilities
python-multipart>=0.0.6
